agent/requirements.txt
# Enigma AI — Agent requirements
# Pinned to versions from the working Enigma venv.

# ── FastAPI server ────────────────────────────────────────────────────────────
fastapi==0.136.0
uvicorn==0.44.0
starlette==1.0.0
python-multipart==0.0.26
sse-starlette==3.3.4

# ── LangChain / LangGraph / LLM ──────────────────────────────────────────────
langchain==1.2.15
langchain-core==1.3.0
langchain-google-genai==4.2.2
langchain-text-splitters==1.1.2
langgraph==1.1.8
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.32

# ── Google Generative AI ──────────────────────────────────────────────────────
google-genai==1.73.1
google-auth==2.49.2

# ── MongoDB ───────────────────────────────────────────────────────────────────
pymongo==4.16.0
dnspython==2.8.0

# ── Data / Validation ─────────────────────────────────────────────────────────
pydantic==2.13.2
pydantic-settings==2.13.1
python-dotenv==1.2.2

# ── HTTP / networking ─────────────────────────────────────────────────────────
httpx==0.28.1
httpcore==1.0.9
httpx-sse==0.4.3
requests==2.33.1
anyio==4.13.0
sniffio==1.3.1
h11==0.16.0

# ── Utilities ─────────────────────────────────────────────────────────────────
tenacity==9.1.4
orjson==3.11.8
PyYAML==6.0.3
packaging==24.2
typing_extensions==4.15.0
annotated-types==0.7.0
jsonpatch==1.33
jsonpointer==3.1.1
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
urllib3==2.6.0
